requirements.txt
pip-bibtexparser